refactor: implement shared zod schema validation for profile forms
