+++ frontend/src/components/PasswordTab.tsx
@@ -9,24 +9,8 @@ import React, {useState} from 'react';
 import Feather from '@expo/vector-icons/Feather';
 import { useForm, Controller } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import * as z from 'zod';
 import {PRIMARY_COLOR} from '../helper/Theme';
-
-const passwordSchema = z.object({
-  old_password: z.string().min(1, 'Old password is required'),
-  new_password: z.string()
-    .min(6, 'At least 6 characters')
-    .regex(/(?=.*[a-z]).{6,}/, 'At least 6 characters with lowercase letter'),
-  confirm_password: z.string().min(1, 'Please confirm your new password'),
-}).refine((data) => data.new_password === data.confirm_password, {
-  message: "Passwords don't match",
-  path: ['confirm_password'],
-}).refine((data) => data.new_password !== data.old_password, {
-  message: "New password must be different from old password",
-  path: ['new_password'],
-});
-
-export type PasswordFormData = z.infer<typeof passwordSchema>;
+import { passwordSchema, PasswordFormData } from '../schemas/profileSchemas';
 
 interface PasswordTabProps {
   handleSubmitPassword: (data: PasswordFormData) => void;
